Refactor files in Customizations UI package
