+++ app/ui-react/packages/ui/src/Customization/apiClientConnectors/ApiConnectorListView.tsx
@@ -1,5 +1,4 @@
 import { Text, Title } from '@patternfly/react-core';
-import * as H from '@syndesis/history';
 import {
   EmptyState,
   ListView,
@@ -7,52 +6,44 @@ import {
   Tooltip,
 } from 'patternfly-react';
 import * as React from 'react';
-import { ButtonLink, PageSection } from '../Layout';
-import { IListViewToolbarProps, ListViewToolbar } from '../Shared';
+import { ButtonLink, PageSection } from '../../Layout';
+import { IListViewToolbarProps, ListViewToolbar } from '../../Shared';
 
-export interface IExtensionListViewProps extends IListViewToolbarProps {
+export interface IApiConnectorListViewProps extends IListViewToolbarProps {
   i18nDescription: string;
   i18nEmptyStateInfo: string;
   i18nEmptyStateTitle: string;
-  i18nLinkImportExtension: H.LocationDescriptor;
-  i18nLinkImportExtensionTip?: H.LocationDescriptor;
+  i18nLinkCreateApiConnector: string;
+  i18nLinkCreateApiConnectorTip?: string;
   i18nName: string;
-  i18nNameFilterPlaceholder: string;
   i18nTitle: string;
-  linkImportExtension: H.LocationDescriptor;
+  linkCreateApiConnector: string;
 }
 
-export class ExtensionListView extends React.Component<
-  IExtensionListViewProps
+export class ApiConnectorListView extends React.Component<
+  IApiConnectorListViewProps
 > {
-  public getImportTooltip() {
-    return (
-      <Tooltip id="importTip">
-        {this.props.i18nLinkImportExtensionTip
-          ? this.props.i18nLinkImportExtensionTip
-          : this.props.i18nLinkImportExtension}
-      </Tooltip>
-    );
-  }
-
   public render() {
     return (
       <PageSection>
         <ListViewToolbar {...this.props}>
           <div className="form-group">
-            <OverlayTrigger overlay={this.getImportTooltip()} placement="top">
+            <OverlayTrigger
+              overlay={this.getCreateConnectorTooltip()}
+              placement="top"
+            >
               <ButtonLink
-                data-testid={'extension-list-view-import-extension'}
-                href={this.props.linkImportExtension}
+                data-testid={'api-connector-list-view-create-api-connector'}
+                href={this.props.linkCreateApiConnector}
                 as={'primary'}
               >
-                {this.props.i18nLinkImportExtension}
+                {this.props.i18nLinkCreateApiConnector}
               </ButtonLink>
             </OverlayTrigger>
           </div>
         </ListViewToolbar>
         {this.props.i18nTitle !== '' && (
-          <Title size="lg">{this.props.i18nTitle}</Title>
+          <Title size="xl">{this.props.i18nTitle}</Title>
         )}
         {this.props.i18nDescription !== '' && (
           <Text
@@ -69,19 +60,32 @@ export class ExtensionListView extends React.Component<
             </EmptyState.Title>
             <EmptyState.Info>{this.props.i18nEmptyStateInfo}</EmptyState.Info>
             <EmptyState.Action>
-              <OverlayTrigger overlay={this.getImportTooltip()} placement="top">
+              <OverlayTrigger
+                overlay={this.getCreateConnectorTooltip()}
+                placement="top"
+              >
                 <ButtonLink
-                  data-testid={'extension-list-view-empty-state-import'}
-                  href={this.props.linkImportExtension}
+                  data-testid={'api-connector-list-view-empty-state-create'}
+                  href={this.props.linkCreateApiConnector}
                   as={'primary'}
                 >
-                  {this.props.i18nLinkImportExtension}
+                  {this.props.i18nLinkCreateApiConnector}
                 </ButtonLink>
               </OverlayTrigger>
             </EmptyState.Action>
           </EmptyState>
         )}
       </PageSection>
+    );
+  }
+
+  private getCreateConnectorTooltip() {
+    return (
+      <Tooltip id="createTip">
+        {this.props.i18nLinkCreateApiConnectorTip
+          ? this.props.i18nLinkCreateApiConnectorTip
+          : this.props.i18nLinkCreateApiConnector}
+      </Tooltip>
     );
   }
 }
