+++ app/ui-react/packages/ui/src/Customization/extensions/ExtensionDetail.css
