+++ app/ui-react/packages/ui/src/Customization/extensions/ExtensionImportReview.tsx
@@ -1,7 +1,7 @@
 import * as H from '@syndesis/history';
 import { Button, Grid } from 'patternfly-react';
 import * as React from 'react';
-import { ButtonLink, Container } from '../Layout';
+import { ButtonLink, Container } from '../../Layout';
 import './ExtensionImportReview.css';
 
 export interface IImportAction {
