+++ app/ui-react/packages/ui/src/Customization/extensions/ExtensionIntegrationsTable.tsx
