+++ app/ui-react/packages/ui/src/Customization/extensions/ExtensionListSkeleton.tsx
