+++ app/ui-react/packages/ui/src/Customization/extensions/ExtensionOverview.css
