+++ app/ui-react/packages/ui/src/Customization/extensions/ExtensionSupports.tsx
