+++ app/ui-react/packages/ui/src/Customization/extensions/index.ts
@@ -1,0 +1,9 @@
+export * from './ExtensionDetail';
+export * from './ExtensionImportCard';
+export * from './ExtensionImportReview';
+export * from './ExtensionIntegrationsTable';
+export * from './ExtensionListItem';
+export * from './ExtensionListSkeleton';
+export * from './ExtensionListView';
+export * from './ExtensionOverview';
+export * from './ExtensionSupports';
