+++ app/ui-react/packages/ui/src/Customization/index.ts
@@ -1,15 +1,2 @@
-export * from './ApiConnectorDetailCard';
-export * from './ApiConnectorDetailsForm';
-export * from './ApiConnectorReview';
-export * from './ApiConnectorListItem';
-export * from './ApiConnectorListSkeleton';
-export * from './ApiConnectorListView';
-export * from './ExtensionDetail';
-export * from './ExtensionImportCard';
-export * from './ExtensionImportReview';
-export * from './ExtensionIntegrationsTable';
-export * from './ExtensionListItem';
-export * from './ExtensionListSkeleton';
-export * from './ExtensionListView';
-export * from './ExtensionOverview';
-export * from './ExtensionSupports';
+export * from './apiClientConnectors';
+export * from './extensions';
